Reorganize UpdateList component into bundle directory
Issue #2807
